master/requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.1.1
django-redis==5.4.0
django-user-agents==0.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.10
kombu==5.4.2
pika==1.3.2
pillow==10.4.0
prompt_toolkit==3.0.48
psycopg2==2.9.9
PyJWT==2.9.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
redis==5.1.1
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
tzdata==2024.2
ua-parser==0.18.0
urllib3==2.2.3
user-agents==2.2.0
vine==5.1.0
wcwidth==0.2.13
